% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBinary.R
\name{pedigraph_binary}
\alias{pedigraph_binary}
\title{file path to be used in a call to pedigraph}
\usage{
pedigraph_binary()
}
\description{
This version checks to make sure it is there and throws an
error with a suggestion of how to get it if it is not there.
}
