% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBinary.R
\name{pedigraph_binary_path}
\alias{pedigraph_binary_path}
\title{return the path where pedigraph should be in the R system paths}
\usage{
pedigraph_binary_path()
}
\description{
return the path where pedigraph should be in the R system paths
}
\keyword{internal}
